+++ lib/fortnite.dart
@@ -4,5 +4,9 @@ export 'resources/fortnite_profile_ids.dart';
 export 'resources/mcp_operations.dart';
 export 'resources/auth_clients.dart';
 
+export 'src/structures/client_options.dart';
+export 'src/structures/device_auth.dart';
+export 'src/structures/http_response.dart';
+
 export 'src/client/client.dart';
 export 'src/client/http.dart';